Scheduled Cleanup Job-Task6/TASK-6/ScheduledCleanupFunc/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python 
# azure-monitor-opentelemetry 

azure-functions
azure-storage-blob
pyodbc
pandas
